[ai-vs-ai] Create a CellsTrait for MovesStore and Board
